Rename module from PID to pid
This avoids shadowing the module with the re-export of the PID class.
Fixes #53.
