+++ tests/test_pid.py
@@ -107,7 +107,7 @@ def test_sample_time():
 
 
 def test_monotonic():
-    from simple_pid.PID import _current_time
+    from simple_pid.pid import _current_time
 
     if sys.version_info < (3, 3):
         assert _current_time == time.time
@@ -134,7 +134,7 @@ def test_auto_mode():
     assert pid(8) == 2
 
     # Last update time should be reset to avoid huge dt
-    from simple_pid.PID import _current_time
+    from simple_pid.pid import _current_time
 
     pid.auto_mode = False
     time.sleep(1)
@@ -159,7 +159,7 @@ def test_separate_components():
 
 
 def test_clamp():
-    from simple_pid.PID import _clamp
+    from simple_pid.pid import _clamp
 
     assert _clamp(None, (None, None)) is None
     assert _clamp(None, (-10, 10)) is None
